Add files via upload
Market Segmentation with Clustering
It is important to define the customer profile. It allows customers to prepare campaigns and deliver them to the right customer. I tried to determine the correct customer profile by using clustering algorithms. I'm new to this. If you have a missing place, don't hesitate to comment. If you have a contribution please upvote!
A company must have different strategies for product promotion for different individuals. Not every individual has the same requirement and demand. So the company has to segment, target and position the product according to the tastes of various individuals.

Segmentation of products based on the revenue generated in different regions is done to understand the market trends.
The segmentation will help the company to devise marketing strategies and promotional schemes to position the right product according to the preference of the consumers in any given segment.